// types/Alumni.type.ts
export interface MembershipPrefillDataType {
  title: string;
  first_name: string;
  last_name: string;
  dob: string;
  category: string;
  nationality: string;
  religion: string;
  address: string;
  pincode: string;
  city: string;
  state: string;
  country: string;
  phone: string;
  alt_phone: string;
  email: string;
  alt_email: string;
  registration_no: string;
  roll_no: string;
  degree: string;
  discipline: string;
  graduation_date: string;
  status: string | null;
}

export interface AlumniMembershipFormSubmissionType {
  membership_level: string;
  sign: File;
}